[Chen Sicong] feat: adding vote search and auto sort ranking
feat: adding vote search
feat: adding auto sort ranking
refactor: change RANKING type

+++ src/com/twu/manageer/TopSearchRankingManager.java
@@ -5,7 +5,8 @@ import com.twu.model.type.TopSearchType;
 import com.twu.model.type.UserType;
 import com.twu.utility.OutputFormatter;
 
-import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Scanner;
 
@@ -13,7 +14,7 @@ public class TopSearchRankingManager {
 
     private final static TopSearchRankingManager INSTANCE = new TopSearchRankingManager();
 
-    private final static List<TopSearch> RANKING = new ArrayList<>();
+    private final static List<TopSearch> RANKING = new LinkedList<>();
 
     private TopSearchRankingManager() {
     }
@@ -33,6 +34,7 @@ public class TopSearchRankingManager {
             System.out.println("4. 撤销热搜");
         } else {
             System.out.println("3. 购买热搜");
+            System.out.println("4. 投票热搜");
         }
         System.out.println("0. 退出登录");
         switch (in.nextInt()) {
@@ -50,7 +52,11 @@ public class TopSearchRankingManager {
                 }
                 break;
             case 4:
-                deleteTopSearch(in);
+                if (currentUserType == UserType.USER_ADMIN) {
+                    deleteTopSearch(in);
+                } else {
+                    voteTopSearch(in);
+                }
                 break;
             case 0:
             default:
@@ -62,9 +68,13 @@ public class TopSearchRankingManager {
     private void checkRanking() {
         OutputFormatter.printInfo("热搜排行榜如下");
         System.out.println(String.format("%-10s %-10s %-10s %-10s", "排名", "名称", "热度", "价格"));
-        for (int i = 0; i < RANKING.size(); i++) {
-            System.out.println(
-                String.format("%-10d %-10s %-10d %-10d", i + 1, RANKING.get(i).getName(), RANKING.get(i).getHeat(), RANKING.get(i).getPrize()));
+        if (RANKING.size() == 0) {
+            OutputFormatter.printInfo("目前热搜为空");
+        } else {
+            for (int i = 0; i < RANKING.size(); i++) {
+                System.out.println(
+                    String.format("%-10d %-10s %-10d %-10d", i + 1, RANKING.get(i).getName(), RANKING.get(i).getHeat(), RANKING.get(i).getPrize()));
+            }
         }
     }
 
@@ -100,6 +110,36 @@ public class TopSearchRankingManager {
 
     private void deleteTopSearch(Scanner in) {
 
+    }
+
+    private void voteTopSearch(Scanner in) {
+        OutputFormatter.printInfo("请输入要投票热搜的排名");
+        int ranking;
+        while (true) {
+            ranking = in.nextInt();
+            if (ranking >= 1 && ranking <= RANKING.size()) {
+                break;
+            } else {
+                OutputFormatter.printError("请输入正确的排名");
+            }
+        }
+        TopSearch votingSearch = RANKING.get(ranking - 1);
+        int haveTicket = UserManager.getCurrentUser().getVoteTicket();
+        OutputFormatter.printInfo(String.format("请输入票数，您现在拥有%d票", haveTicket));
+        int tickets;
+        while (true) {
+            tickets = in.nextInt();
+            if (tickets >= 1 && tickets <= haveTicket) {
+                break;
+            } else {
+                OutputFormatter.printError("请输入正确的票数");
+            }
+        }
+
+        votingSearch.setHeat(votingSearch.getHeat() + (votingSearch.getTopSearchType() == TopSearchType.SEARCH_SUPER
+            ? tickets * 2 : tickets));
+
+        Collections.sort(RANKING);
     }
 
 }
